cleanup(generic-snmp): lib renamed into snmp
