+++ experimental/Cargo.toml
@@ -1,13 +1,6 @@
 [package]
-name = "conn"
-version = "1.1.0"
-
-[lib]
-crate-type = ["cdylib"]
-
-[[bin]]
 name = "generic-snmp"
-path = "src/main.rs"
+version = "1.0.0"
 
 [build-dependencies]
 lalrpop = "0.22.1"
@@ -22,4 +15,3 @@ lexopt = "0.3.0"
 serde = { version = "1.0.217", features = ["derive"] }
 serde_json = "1.0.138"
 lalrpop-util = { version = "0.22.1", features = ["lexer"] }
-
